# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIC 2026 Retrieval System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingestions": {
      "get": {
        "tags": [
          "Ingestion"
        ],
        "summary": "List Ingestion Jobs",
        "operationId": "list_ingestion_jobs_api_v1_ingestions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Create Ingestion Job",
        "operationId": "create_ingestion_job_api_v1_ingestions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngestionJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIngestionJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestions/{job_id}": {
      "get": {
        "tags": [
          "Ingestion"
        ],
        "summary": "Get Ingestion Job",
        "operationId": "get_ingestion_job_api_v1_ingestions__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateIngestionJobRequest": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/IngestionEntity"
          },
          "manifest_path": {
            "type": "string",
            "minLength": 1,
            "title": "Manifest Path"
          },
          "collection_name": {
            "type": "string",
            "minLength": 1,
            "title": "Collection Name"
          },
          "feature_profile": {
            "type": "string",
            "minLength": 1,
            "title": "Feature Profile"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "manifest_path",
          "collection_name",
          "feature_profile"
        ],
        "title": "CreateIngestionJobRequest"
      },
      "CreateIngestionJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionStatus",
            "default": "queued"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "collection_name"
        ],
        "title": "CreateIngestionJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestionEntity": {
        "type": "string",
        "enum": [
          "frames",
          "clips"
        ],
        "title": "IngestionEntity"
      },
...
      "IngestionJobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/IngestionJobStatusResponse"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "jobs"
        ],
        "title": "IngestionJobListResponse"
      },
      "IngestionJobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionStatus"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestionStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestionProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "collection_name"
        ],
        "title": "IngestionJobStatusResponse"
      },
      "IngestionProgress": {
        "properties": {
          "completed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Completed"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "percent": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent"
          }
        },
        "type": "object",
        "required": [
          "completed",
          "total",
          "percent"
        ],
        "title": "IngestionProgress"
      },
      "IngestionStage": {
        "type": "string",
        "enum": [
          "validating",
          "creating_collection",
          "creating_payload_indexes",
          "upserting",
          "optimizing",
          "completed"
        ],
        "title": "IngestionStage"
      },
      "IngestionStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "IngestionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}